requirements.txt
fastembed==0.3.6
openai==1.45.0
python-dotenv==1.0.1
qdrant-client==1.11.1
retry==0.9.2
rich==13.8.1
spacy==3.7.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sqlglot==25.21.1
sqlparse==0.5.1
tqdm==4.66.5
func_timeout==4.3.5
colorlog==6.8.2
chardet==5.2.0
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
kaleido==0.2.1